changed util file name, moved map outside jsx
